Create 2018_10_09_01_fixed_eye_of_azshara.sql
The loot of 90% of the creatures in the Ojo de Azshara was updated.
The revive inside the dungeon was updated when any player dies.
The 4 npc rare elites that did not appear in the dungeon were added:

Shellmaw
Jaggen Ra
Gom Crabbar
Dread Captain Thedon

The quest 38286 was updated
https://www.wowhead.com/quest=38286/eye-of-azshara-wrath-of-azshara

The dungeon is almost still 95% Blizzard. 
The dungeon needs a little more work.
The dungeon needs the object to be added in order to put the mythical stone.